Add support for weak refs, remove rc_free() and rc_count() as no longer needed.
This adds support for weak references.   That removes the original need for an explicit call
to rc_free in client code (just use a weak reference) and the need to check rc_count() as
well.

Code that uses weak refs must remember to call rc_deref() to make sure that the reference is good
before use!