Simplified the API.  Removed rc_deref().
There was no real reason for rc_deref().  If you want some object to stay
around while you are modifying/reading it, then you need to acquire a strong
reference on it.   So, 99% or 100% of the use cases for rc_deref() are
replaceable with rc_inc/rc_dec.   And, without having taken a strong ref
to the data, you cannot be sure that something else would not have dropped
the last reference.