Working commit.   Yet another RC API.  Data structure changes.
This is the first commit of the next RC API.   References are handled with hash tables.

Vectors and hashtables no longer care about RC.   They do not use it.

RC has been changed to allow special cleanup handlers to be added to the clean up process.

The base util code compiles, but lib.c is very broken.