AB Logix starting to compile all the way.
This does not work.   It does compile.   Most of the code to set up
a session is now written.   Changes were made to bytebuf and to
the platform code to support extra functionality.

The logix code was refactored into a new subdirectory under ab.

Jobs now replace protothreads.   There is now a job thread pool.
All completely untested.