First cut at new structure with only generic tags and implementation PLCs.
This has the first cut at the new way of doing things.   The generic tag is the
only visible tag type.   Each PLC implementation must provide tag_abort, tag_read,
tag_status, and tag_write functions.   If there is additional state that is needed,
the implementation must handle it.

Jobs were removed as were async socket.   The rc_thread service was created to allow
threads to be handled via RC.

None of this is tested.