Removed reaper thread etc.   Now when a rc_thread is destroyed, the thread is joined.
The previous API let the thread continue to run even though the data it pointed to might
have been deleted.   Now, when a thread is destroyed via rc_dec, the caller waits until
the thread can be joined.  Not ideal and could block but better than reading and writing
unallocated memory!